Only listen mousemove once
初始化时监听, 共享同一个 mousePosition
